pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-central"
version = "0.1.0"
description = "AI-Ops CLI: Orchestrating the Agency"
dependencies = [
    "typer",
    "rich",
    "gitpython",
    "pyyaml",
    "pydantic",
    "sentence-transformers"
]

[project.scripts]
ai = "agent_central.main:app"
